Added Mockito testing

+++ QuizApp-Microservices/question-service/src/main/java/com/munaf/question_service/repositories/QuestionsRepository.java
@@ -12,5 +12,5 @@ public interface QuestionsRepository extends JpaRepository<QuestionEntity, Long>
     List<QuestionEntity> findByCategory(String category);
 
     @Query(value = "SELECT q.id FROM questions_table q WHERE q.category=?1 ORDER BY RAND() LIMIT ?2", nativeQuery = true)
-    List<Integer> getRandomQuestionsByCategoryWithSize(String category, Integer size);
+    List<Long> getRandomQuestionsByCategoryWithSize(String category, Integer size);
 }
